requirements.txt
chardet==3.0.4
Click==7.0
Flask==0.12
Flask-Login==0.5.0
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
glog==0.3.1
gunicorn==19.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
mongoengine==0.19.1
numpy==1.16.6
opencv-python==4.2.0.32
pandas==0.24.2
Pillow==6.2.2
pymongo==3.10.1
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2019.3
requests==2.22.0
six==1.14.0
tqdm==4.42.1
urllib3==1.25.8
uWSGI==2.0.18
Werkzeug==1.0.0
WTForms==2.2.1
